+++ src/profile/map/map.cpp
@@ -192,8 +192,19 @@ void map::read_data(toml::value &data)
             m_data_type = map_data_type::note;
 
             LOG_DEBUG << " --> Line " << data.location().line() << " :: Parameter '" << CFG_KEY_NOTE << "' = '"
-                      << m_data
-                      << "'" << LOG_END
+                      << m_data << "'" << LOG_END
+        } else if (data.contains(CFG_KEY_PITCH_BEND)) {
+            m_data = static_cast<unsigned char>( data[CFG_KEY_PITCH_BEND].as_integer());
+            m_data_type = map_data_type::note;
+
+            LOG_DEBUG << " --> Line " << data.location().line() << " :: Parameter '" << CFG_KEY_PITCH_BEND << "' = '"
+                      << m_data << "'" << LOG_END
+        } else if (data.contains(CFG_KEY_PROGRAM_CHANGE)) {
+            m_data = static_cast<unsigned char>( data[CFG_KEY_PROGRAM_CHANGE].as_integer());
+            m_data_type = map_data_type::note;
+
+            LOG_DEBUG << " --> Line " << data.location().line() << " :: Parameter '" << CFG_KEY_PROGRAM_CHANGE
+                      << "' = '" << m_data << "'" << LOG_END
         } else {
             LOG_ERROR << " --> Line " << data.location().line() << " :: Parameter '" << CFG_KEY_CC << "' or "
                       << "Paramater '" << CFG_KEY_NOTE << "' is missing"